+++ src/constants/symbols.ts
@@ -1,0 +1,2 @@
+// tslint:disable-next-line:variable-name
+export const AttributeMetadata = Symbol('AttributeMetadata');