+++ src/interfaces/attribute-decorator-options.interface.ts
@@ -1,0 +1,6 @@
+import { PropertyConverter } from './property-converter.interface';
+
+export interface AttributeDecoratorOptions {
+  serializedName?: string;
+  converter?: PropertyConverter;
+}