+++ src/interfaces/overrides.interface.ts
@@ -1,0 +1,4 @@
+export interface Overrides {
+  getDirtyAttributes?: (attributedMetadata: any) => object;
+  toQueryString?: (params: any) => string;
+}