+++ src/services/json-api-datastore.service.ts
@@ -13,6 +13,7 @@ import { JsonApiQueryData } from '../models/json-api-query-data';
 import * as qs from 'qs';
 import { DatastoreConfig } from '../interfaces/datastore-config.interface';
 import { ModelConfig } from '../interfaces/model-config.interface';
+import { AttributeMetadata } from '../constants/symbols';
 
 export type ModelType<T extends JsonApiModel> = { new(datastore: JsonApiDatastore, data: any): T; };
 
@@ -21,7 +22,11 @@ export class JsonApiDatastore {
   // tslint:disable:variable-name
   private _headers: Headers;
   private _store: {[type: string]: {[id: string]: JsonApiModel}} = {};
-
+  // tslint:disable:max-line-length
+  private getDirtyAttributes: Function = this.datastoreConfig.overrides && this.datastoreConfig.overrides.getDirtyAttributes ? this.datastoreConfig.overrides.getDirtyAttributes : this._getDirtyAttributes;
+  private toQueryString: Function = this.datastoreConfig.overrides && this.datastoreConfig.overrides.toQueryString ? this.datastoreConfig.overrides.toQueryString : this._toQueryString;
+  // tslint:enable:max-line-length
+  
   protected config: DatastoreConfig;
 
   constructor(private http: Http) {}
@@ -73,7 +78,7 @@ export class JsonApiDatastore {
     return new modelType(this, { attributes: data });
   }
 
-  private getDirtyAttributes(attributesMetadata: any): { string: any} {
+  private _getDirtyAttributes(attributesMetadata: any): { string: any} {
     const dirtyData: any = {};
 
     for (const propertyName in attributesMetadata) {
@@ -191,13 +196,21 @@ export class JsonApiDatastore {
       if (data.hasOwnProperty(key)) {
         if (data[key] instanceof JsonApiModel) {
           relationships = relationships || {};
-          relationships[key] = {
-            data: this.buildSingleRelationshipData(data[key])
-          };
+          
+          if (data[key].id) {
+            relationships[key] = {
+              data: this.buildSingleRelationshipData(data[key])
+            };
+          }
         } else if (data[key] instanceof Array && data[key].length > 0 && this.isValidToManyRelation(data[key])) {
           relationships = relationships || {};
+
+          const relationshipData = data[key]
+            .filter((model: JsonApiModel) => model.id)
+            .map((model: JsonApiModel) => this.buildSingleRelationshipData(model));
+
           relationships[key] = {
-            data: data[key].map((model: JsonApiModel) => this.buildSingleRelationshipData(model))
+            data: relationshipData
           };
         }
       }
@@ -305,15 +318,7 @@ export class JsonApiDatastore {
 
   protected parseMeta(body: any, modelType: ModelType<JsonApiModel>): any {
     const metaModel: any = Reflect.getMetadata('JsonApiModelConfig', modelType).meta;
-    const jsonApiMeta = new metaModel();
-
-    for (const key in body) {
-      if (jsonApiMeta.hasOwnProperty(key)) {
-        jsonApiMeta[key] = body[key];
-      }
-    }
-
-    return jsonApiMeta;
+    return new metaModel(body);
   }
 
   private getOptions(customHeaders?: Headers): RequestOptions {
@@ -340,7 +345,7 @@ export class JsonApiDatastore {
     return new RequestOptions({ headers: requestHeaders });
   }
 
-  private toQueryString(params: any) {
+  private _toQueryString(params: any): string {
     return qs.stringify(params, { arrayFormat: 'brackets' });
   }
 
@@ -361,7 +366,7 @@ export class JsonApiDatastore {
   private resetMetadataAttributes<T extends JsonApiModel>(res: T, attributesMetadata: any, modelType: ModelType<T>) {
     // TODO check why is attributesMetadata from the arguments never used
     // tslint:disable-next-line:no-param-reassign
-    attributesMetadata = Reflect.getMetadata('Attribute', res);
+    attributesMetadata = res[AttributeMetadata];
 
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
@@ -373,7 +378,7 @@ export class JsonApiDatastore {
       }
     }
 
-    Reflect.defineMetadata('Attribute', attributesMetadata, res);
+    res[AttributeMetadata] = attributesMetadata;
     return res;
   }
 
@@ -389,8 +394,14 @@ export class JsonApiDatastore {
         });
 
         if (propertyHasMany) {
-          if (relationshipModel[propertyHasMany.propertyName] !== undefined) {
+          relationshipModel[propertyHasMany.propertyName] = relationshipModel[propertyHasMany.propertyName] || [];
+          
+          const indexOfModel = relationshipModel[propertyHasMany.propertyName].indexOf(model);
+
+          if (indexOfModel === -1) {
             relationshipModel[propertyHasMany.propertyName].push(model);
+          } else {
+            relationshipModel[propertyHasMany.propertyName][indexOfModel] = model;
           }
         }
       }
@@ -407,7 +418,7 @@ export class JsonApiDatastore {
   private transformSerializedNamesToPropertyNames<T extends JsonApiModel>(modelType: ModelType<T>, attributes: any) {
     const serializedNameToPropertyName = this.getModelPropertyNames(modelType.prototype);
     const properties: any = {};
-
+    
     Object.keys(serializedNameToPropertyName).forEach((serializedName) => {
       if (attributes[serializedName]) {
         properties[serializedNameToPropertyName[serializedName]] = attributes[serializedName];